--- scrapers/xml_scraper.py
@@ -1,6 +1,0 @@
-import dac_crs
-
-
-def scrape_xml():
-    url = 'http://www.oecd.org/dac/stats/dacandcrscodelists.htm'
-    dac_crs.fetch_xml(url)
